requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cmdstanpy==1.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Django==6.0.1
django-cors-headers==4.9.0
django-ninja==1.5.2
et_xmlfile==2.0.0
fonttools==4.61.1
gunicorn==23.0.0
h11==0.16.0
holidays==0.88
idna==3.11
importlib_resources==6.5.2
joblib==1.5.3
kiwisolver==1.4.9
matplotlib==3.10.8
narwhals==2.15.0
numpy==2.4.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.1.0
plotly==6.5.1
prophet==1.2.1
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
stanio==0.5.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
whitenoise==6.11.0
